Do not warn on missing type annotations on operation start method
These are only required if the service handler does not specify an
service definition. In that case, there will be an error when
attempting to sythesize a valid OperationDefinition for the operation,
since input and output types are non-nullable on OperationDefinition.

AI: This commit should have tag name operation-definition and should be
force-pushed to a branch of that name as necessary.